\name{LIST}
\alias{LIST}
\title{
List BASIC Programs
}
\description{
Lists whichever \acronym{BASIC} program is loaded.
}
\usage{
LIST()
}
\value{
Returns a character vector, containing the complete listing of whichever
\acronym{BASIC} program is in memory at the time.
}
\seealso{
\code{\link{NotRun}}
}
\examples{
# List the current BASIC program.
LIST()
}
